refactor(desktop): verified redundancy pass over the interaction PR
Dead weight, each verified consumer-less before cutting:

- emptyMessage (both lists), PanelTransition/DropRegion className:
  no caller ever passed them.
- BulkPanelItem/BulkSourceContext/SourceMember exports had no
  importer; scoped back.
- RightPanel's detail/edit selectedKey payload had zero readers —
  the variants are plain tags now.
- use-resource-groups' useCallbacks were compiler-covered ceremony.

Repetition at the rule-of-three threshold:

- matrixGroup() builds the coverage-matrix rows once (3 sites).
- MultiSelectToggle replaces three inline toolbar blocks — two of
  which were hand-rolled div[role=button] keyboard handling that a
  plain HeroUI icon Button does properly.
- (The create-group submit flow shows up only twice — left inline.)

skill-list drops from 699 to 363 lines: its pure derivation pipeline
(availability filter → lock clustering → custom sections → the loose
order → shift-range entries, plus the expansion state that folds
collapsed sections into single entries) moves to use-skill-sections;
the component keeps wiring and rendering only. Behavior unchanged —
61/61 e2e.
